store: connect to web socket (#6)
This commit adds support for the new web socket of a store.

It removes the `didAppend` event of an agent and replaces it with `didSave`, which is called everytime a segment is saved with the segment as an argument.

Unfortunately I couldn't find an easy way to test web sockets so there isn't a new test case.