add multi-process support
The process name is now added to the segment before it is sent to the
store. It is saved in segment.link.meta.process.
An agent can be instanciated as follows: var agent = Agent.create({url})
Processes can be added to this agent through its instance using
the agent.addProcess() method. Multiple processes may, or may not share
a store and/or a fossilizer. The http server provided by the
agent allows to list and delete processes, create new instances (maps)
for a given process and add/delete segments of a map.
Note that it is not possible to query a store for map IDs or segments
across all processes, it must me specified.
All the processes' information are rendered on the 'GET "/"' route of
the http server.
Add support for es8 features : new babel plugins were added to insure
the compatibility with older versions of node.