agent-js: update PATH with npm binaries in Docker
To optimize the caching performed by Docker, we first copy the
package.json to the Docker image and install the dependencies. We then
use a mounted volume for all the src files and the actual package.json
from which we run the npm commands. Since this package.json is not at
the same level where the dependencies were installed, it isn't aware of
the available binaries. We use to install these dependencies globally to
work around this issue, but it is slower at build time and adds
duplication.
We chose instead to update the PATH so that is takes into account the
"local" dependencies.