agent: introduce a client factory for client caching (#283)
The way we used to cache the clients created was not compatible with
having mutliple client implementation. For instance, the agent object
should not depend directly on a client implementation as it used to be
the case. This forces the user of the Agent to use a factory to
instantiate its clients though (as seen in the README).